refactor: 모각 service 클래스 interface화(#2)
